workflow/requirements.txt
boto3==1.42.3
botocore==1.42.3
jmespath==1.0.1
lockfile==0.12.2
luigi==3.6.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.16.0
six==1.17.0
tenacity==8.5.0
tornado==6.5.2
urllib3==2.5.0
